+++ CHANGELOG.md
@@ -26,6 +26,7 @@
 * Attempt to set tid of recovered threads using clone3
 * Track opened files that were deleted so we can map them in state loading
 * Add versions for SteamUser API
+* Add dummy hooks for pulseaudio
 
 ### Changed